When no rmf_cat use IMP.rmf
Now extract_models_single_rmf works if rmf_cat is not
available.  Uses subprocess to test for rmf_cat binary.

Also, a path bug was fixed:
the full RMF file path relative to the model directory
is contained in the rmf3_file field in `selected_models....dat`.
The `traj` field does not need to be added.